+++ Sources/Aida128CLI/main.swift
@@ -4,13 +4,14 @@ import Foundation
 
 @main
 struct Aida128CLI {
-    static func main() {
+    static func main() async {
         do {
             switch try CommandLineParser.parse(Array(CommandLine.arguments.dropFirst())) {
             case .help:
                 print(CommandLineParser.usage)
             case let .run(options):
-                let report = try BenchmarkRunner.run(configuration: options.benchmarkConfiguration)
+                let nativeReport = try BenchmarkRunner.run(configuration: options.benchmarkConfiguration)
+                let report = await BenchmarkRunner.enriched(nativeReport)
                 let data = try ReportRenderer.render(report, format: options.format)
                 if let outputPath = options.outputPath {
                     try data.write(to: URL(fileURLWithPath: outputPath), options: .atomic)